backend/python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.18",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.30",
    "langchain-pinecone>=0.2.11",
    "langchain-text-splitters>=0.3.9",
    "langgraph>=0.6.5",
    "openai>=1.99.9",
    "pinecone>=7.3.0",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
